spooky noises
split some stuff in the db
added a simpler way to unlock / lock tabs
